+++ src/main/java/com/sandymandy/pleasurecraft/entity/ai/goal/StripGoal.java
@@ -41,7 +41,7 @@ public class StripGoal extends Goal {
         if(!girl.hasStripAnim()) return;
         if(started) {
             if (!girl.isFrozenInPlace()) girl.setFreeze(true);
-            if (girl.getAnimationKeyFrameEvent().equals("becomeNude".toLowerCase()) && !stripTrigged) {
+            if (girl.getAnimationKeyFrameEvent().contains("becomeNude".toLowerCase()) && !stripTrigged) {
                 girl.setStripped(!girl.isStripped()); // toggle stripped state
                 stripTrigged = true;
             }